refactor user, therapist

+++ src/main/java/vn/vinhdeptrai/skincarebookingsystem/service/TherapistService.java
@@ -35,6 +35,20 @@ public class TherapistService {
     PasswordEncoder passwordEncoder;
     CloudinaryUtil cloudinaryUtil;
     RoleRepository roleRepository;
+
+    public List<TherapistResponse> getAll() {
+        List<Therapist> therapists = therapistRepository.findAll();
+        if(therapists.isEmpty()){
+            throw new AppException(ErrorCode.THERAPIST_NOT_FOUND);
+        }
+        return therapists.stream().map(therapistMapper::toTherapistResponse).collect(Collectors.toList());
+    }
+    public TherapistResponse getById(Integer id) {
+        Therapist therapist = therapistRepository.findById(id).orElseThrow(
+                () -> new AppException(ErrorCode.THERAPIST_NOT_FOUND)
+        );
+        return therapistMapper.toTherapistResponse(therapist);
+    }
     public TherapistResponse create(TherapistRequest therapistRequest, MultipartFile file) throws IOException {
         if(userRepository.existsByUsername(therapistRequest.getUsername())){
             throw new AppException(ErrorCode.USER_EXISTED);
@@ -42,7 +56,7 @@ public class TherapistService {
         Role therapistRole = roleRepository.findByName(PredefinedRole.THERAPIST_ROLE).orElseThrow(
                 () -> new AppException(ErrorCode.ROLE_NOT_FOUND));
         Therapist therapist = therapistMapper.toTherapist(therapistRequest);
-        therapist.setPassword(passwordEncoder.encode(therapistRequest.getPassword()));
+//        therapist.setPassword(passwordEncoder.encode(therapistRequest.getPassword()));
         therapist.setImage(cloudinaryUtil.uploadImage(file));
         therapist.setRole(Set.of(therapistRole));
 
@@ -57,16 +71,19 @@ public class TherapistService {
         }
         Role therapistRole = roleRepository.findByName(PredefinedRole.THERAPIST_ROLE).orElseThrow(
                 () -> new AppException(ErrorCode.ROLE_NOT_FOUND));
-       therapistMapper.updateTherapist(therapist,therapistRequest);
-        therapist.setPassword(passwordEncoder.encode(therapistRequest.getPassword()));
-        therapist.setImage(cloudinaryUtil.uploadImage(file));
+        therapistMapper.updateTherapist(therapist,therapistRequest);
+//        therapist.setPassword(passwordEncoder.encode(therapistRequest.getPassword()));
+        if(file != null){
+            therapist.setImage(cloudinaryUtil.uploadImage(file));
+        }
         return therapistMapper.toTherapistResponse(therapistRepository.save(therapist));
     }
     public void delete(int id) throws AppException {
+        if(userRepository.existsById(id)){
+            throw new AppException(ErrorCode.THERAPIST_NOT_FOUND);
+        }
         therapistRepository.deleteById(id);
     }
 
-    public List<TherapistResponse> getAll() {
-        return therapistRepository.findAll().stream().map(therapistMapper::toTherapistResponse).collect(Collectors.toList());
-    }
+
 }
